backend/requirements.txt
fastapi
uvicorn[standard]
openai
python-dotenv
azure-data-tables
azure-identity
pydantic
requests 